MLPart/requirements.txt
beautifulsoup4==4.14.2
nltk==3.9.2
grpcio==1.76.0
spacy==3.8.11
openai==2.8.1
langchain-core==1.0.7
langchain-experimental==0.4.0
langchain-text-splitters==1.0.0
qdrant-client==1.16.0
neo4j==6.0.3
python-dotenv==1.2.1
fastapi
uvicorn
